Consolidate file filter classes into single file
