+++ src/main/scala/dpla/ingestion3/harvesters/file/HathiFileHarvester.scala
@@ -2,12 +2,12 @@ package dpla.ingestion3.harvesters.file
 
 import java.io.{File, FileInputStream}
 import java.util.zip.GZIPInputStream
-
 import dpla.ingestion3.confs.i3Conf
 import org.apache.commons.io.IOUtils
 import org.apache.log4j.Logger
 import org.apache.spark.sql.{DataFrame, SparkSession}
 import com.databricks.spark.avro._
+import dpla.ingestion3.harvesters.file.FileFilters.GzFileFilter
 import dpla.ingestion3.mappers.utils.XmlExtractor
 import org.apache.tools.tar.TarInputStream
