+++ src/main/scala/dpla/ingestion3/harvesters/file/LcCsvFileHarvester.scala
@@ -1,9 +1,9 @@
 package dpla.ingestion3.harvesters.file
 
 import java.io.File
-
 import com.databricks.spark.avro._
 import dpla.ingestion3.confs.i3Conf
+import dpla.ingestion3.harvesters.file.FileFilters.CsvFileFilter
 import dpla.ingestion3.mappers.utils.JsonExtractor
 import org.apache.log4j.Logger
 import org.apache.spark.sql.{DataFrame, SparkSession}
