+++ src/main/scala/dpla/ingestion3/harvesters/file/MissouriFileHarvester.scala
@@ -2,7 +2,6 @@ package dpla.ingestion3.harvesters.file
 
 import java.io.{BufferedReader, File, FileInputStream, InputStreamReader}
 import java.util.zip.ZipInputStream
-
 import dpla.ingestion3.confs.i3Conf
 import dpla.ingestion3.mappers.utils.JsonExtractor
 import org.apache.commons.io.IOUtils
@@ -12,6 +11,7 @@ import org.apache.spark.sql.functions._
 import org.json4s.jackson.JsonMethods._
 import org.json4s.{JValue, _}
 import com.databricks.spark.avro._
+import dpla.ingestion3.harvesters.file.FileFilters.ZipFileFilter
 
 import scala.util.{Failure, Success, Try}
