+++ src/main/scala/dpla/ingestion3/harvesters/file/NaraDeltaHarvester.scala
@@ -2,10 +2,10 @@ package dpla.ingestion3.harvesters.file
 
 import java.io.{BufferedReader, File, FileInputStream}
 import java.util.zip.GZIPInputStream
-
 import com.databricks.spark.avro._
 import dpla.ingestion3.confs.i3Conf
 import dpla.ingestion3.dataStorage.InputHelper
+import dpla.ingestion3.harvesters.file.FileFilters.GzFileFilter
 import dpla.ingestion3.harvesters.{AvroHelper, Harvester}
 import dpla.ingestion3.utils.{FlatFileIO, Utils}
 import org.apache.avro.Schema
