+++ src/main/scala/dpla/ingestion3/harvesters/file/NaraFileHarvester.scala
@@ -2,9 +2,9 @@ package dpla.ingestion3.harvesters.file
 
 import java.io.{BufferedReader, File, FileInputStream}
 import java.util.zip.GZIPInputStream
-
 import com.databricks.spark.avro._
 import dpla.ingestion3.confs.i3Conf
+import dpla.ingestion3.harvesters.file.FileFilters.{GzFileFilter, XmlFileFilter}
 import dpla.ingestion3.harvesters.{AvroHelper, Harvester}
 import dpla.ingestion3.utils.{FlatFileIO, Utils}
 import org.apache.avro.Schema
