--- src/main/scala/dpla/ingestion3/harvesters/file/XmlFileFilter.scala
@@ -1,7 +1,0 @@
-package dpla.ingestion3.harvesters.file
-
-import java.io.{File, FileFilter}
-
-class XmlFileFilter extends FileFilter {
-  override def accept(pathname: File): Boolean = pathname.getName.endsWith("xml")
-}
